requirements.txt
pandas
yfinance
streamlit
pyyaml
